requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
